Reorganize contract test module, and fixtures, to save on gas.
