+++ tests/blockchain/eth/contracts/main/miners_escrow/test_miners_escrow_operations.py
@@ -1,0 +1,89 @@
+import pytest
+from eth_tester.exceptions import TransactionFailed
+from web3.contract import Contract
+
+
+VALUE_FIELD = 0
+DECIMALS_FIELD = 1
+CONFIRMED_PERIOD_1_FIELD = 2
+CONFIRMED_PERIOD_2_FIELD = 3
+LAST_ACTIVE_PERIOD_FIELD = 4
+
+
+@pytest.mark.slow
+def test_verifying_state(testerchain, token):
+    creator = testerchain.interface.w3.eth.accounts[0]
+    miner = testerchain.interface.w3.eth.accounts[1]
+
+    # Deploy contract
+    contract_library_v1, _ = testerchain.interface.deploy_contract(
+        'MinersEscrow', token.address, 1, int(8e7), 4, 4, 2, 100, 1500
+    )
+    dispatcher, _ = testerchain.interface.deploy_contract('Dispatcher', contract_library_v1.address)
+
+    # Deploy second version of the contract
+    contract_library_v2, _ = testerchain.interface.deploy_contract(
+        'MinersEscrowV2Mock', token.address, 2, 2, 2, 2, 2, 2, 2, 2
+    )
+
+    contract = testerchain.interface.w3.eth.contract(
+        abi=contract_library_v2.abi,
+        address=dispatcher.address,
+        ContractFactoryClass=Contract)
+    assert 1500 == contract.functions.maxAllowableLockedTokens().call()
+
+    # Initialize contract and miner
+    policy_manager, _ = testerchain.interface.deploy_contract(
+        'PolicyManagerForMinersEscrowMock', token.address, contract.address
+    )
+    tx = contract.functions.setPolicyManager(policy_manager.address).transact()
+    testerchain.wait_for_receipt(tx)
+
+    tx = contract.functions.initialize().transact({'from': creator})
+    testerchain.wait_for_receipt(tx)
+    tx = token.functions.transfer(miner, 1000).transact({'from': creator})
+    testerchain.wait_for_receipt(tx)
+    balance = token.functions.balanceOf(miner).call()
+    tx = token.functions.approve(contract.address, balance).transact({'from': miner})
+    testerchain.wait_for_receipt(tx)
+    tx = contract.functions.deposit(balance, 1000).transact({'from': miner})
+    testerchain.wait_for_receipt(tx)
+
+    # Upgrade to the second version
+    tx = dispatcher.functions.upgrade(contract_library_v2.address).transact({'from': creator})
+
+    testerchain.wait_for_receipt(tx)
+    assert contract_library_v2.address == dispatcher.functions.target().call()
+    assert 1500 == contract.functions.maxAllowableLockedTokens().call()
+    assert policy_manager.address == contract.functions.policyManager().call()
+    assert 2 == contract.functions.valueToCheck().call()
+    tx = contract.functions.setValueToCheck(3).transact({'from': creator})
+    testerchain.wait_for_receipt(tx)
+    assert 3 == contract.functions.valueToCheck().call()
+
+    # Can't upgrade to the previous version or to the bad version
+    contract_library_bad, _ = testerchain.interface.deploy_contract(
+        'MinersEscrowBad', token.address, 2, 2, 2, 2, 2, 2, 2
+    )
+
+    with pytest.raises((TransactionFailed, ValueError)):
+        tx = dispatcher.functions.upgrade(contract_library_v1.address).transact({'from': creator})
+        testerchain.wait_for_receipt(tx)
+    with pytest.raises((TransactionFailed, ValueError)):
+        tx = dispatcher.functions.upgrade(contract_library_bad.address).transact({'from': creator})
+        testerchain.wait_for_receipt(tx)
+
+    # But can rollback
+    tx = dispatcher.functions.rollback().transact({'from': creator})
+    testerchain.wait_for_receipt(tx)
+    assert contract_library_v1.address == dispatcher.functions.target().call()
+    assert policy_manager.address == contract.functions.policyManager().call()
+
+    with pytest.raises((TransactionFailed, ValueError)):
+        tx = contract.functions.setValueToCheck(2).transact({'from': creator})
+        testerchain.wait_for_receipt(tx)
+
+    # Try to upgrade to the bad version
+    with pytest.raises((TransactionFailed, ValueError)):
+        tx = dispatcher.functions.upgrade(contract_library_bad.address).transact({'from': creator})
+        testerchain.wait_for_receipt(tx)
